requirements.txt
torch
transformers
clip
duckduckgo-search
aiohttp
pillow
matplotlib
pyyaml
gradio
google-generativeai
python-dotenv
git+https://github.com/openai/CLIP.git